refactor: split QueryTab.swift (773 lines) into 4 focused files
- QueryTabState.swift (210 lines) — TabType, PersistedTab, TabPendingChanges,
  SortDirection, SortColumn, SortState, PaginationState, ColumnLayoutState
- RowBuffer.swift (75 lines) — reference-type wrapper for large result data
- QueryTabManager.swift (204 lines) — @Observable tab collection manager
- QueryTab.swift (305 lines) — just the QueryTab struct (was 773 lines)

No type renames, no API changes, no import changes in consuming files.
