+++ TablePro/Models/Query/QueryTab.swift
@@ -10,277 +10,6 @@ import Observation
 import os
 import TableProPluginKit
 
-/// Type of tab
-enum TabType: Equatable, Codable, Hashable {
-    case query       // SQL editor tab
-    case table       // Direct table view tab
-    case createTable // Create new table tab
-}
-
-/// Minimal representation of a tab for persistence
-struct PersistedTab: Codable {
-    let id: UUID
-    let title: String
-    let query: String
-    let tabType: TabType
-    let tableName: String?
-    var isView: Bool = false
-    var databaseName: String = ""
-    var schemaName: String?
-    var sourceFileURL: URL?
-}
-
-/// Stores pending changes for a tab (used to preserve state when switching tabs)
-struct TabPendingChanges: Equatable {
-    var changes: [RowChange]
-    var deletedRowIndices: Set<Int>
-    var insertedRowIndices: Set<Int>
-    var modifiedCells: [Int: Set<Int>]
-    var insertedRowData: [Int: [String?]]  // Lazy storage for inserted row values
-    var primaryKeyColumn: String?
-    var columns: [String]
-
-    init() {
-        self.changes = []
-        self.deletedRowIndices = []
-        self.insertedRowIndices = []
-        self.modifiedCells = [:]
-        self.insertedRowData = [:]
-        self.primaryKeyColumn = nil
-        self.columns = []
-    }
-
-    var hasChanges: Bool {
-        !changes.isEmpty || !insertedRowIndices.isEmpty || !deletedRowIndices.isEmpty
-    }
-}
-
-/// Sort direction for column sorting
-enum SortDirection: Equatable {
-    case ascending
-    case descending
-
-    var indicator: String {
-        switch self {
-        case .ascending: return "▲"
-        case .descending: return "▼"
-        }
-    }
-
-    mutating func toggle() {
-        self = self == .ascending ? .descending : .ascending
-    }
-}
-
-/// A single column in a multi-column sort
-struct SortColumn: Equatable {
-    var columnIndex: Int
-    var direction: SortDirection
-}
-
-/// Tracks sorting state for a table (supports multi-column sort)
-struct SortState: Equatable {
-    var columns: [SortColumn] = []
-
-    init() {}
-
-    var isSorting: Bool { !columns.isEmpty }
-
-    // Backward-compatible computed properties for single-column access
-    var columnIndex: Int? { columns.first?.columnIndex }
-    var direction: SortDirection { columns.first?.direction ?? .ascending }
-}
-
-/// Tracks pagination state for navigating large datasets
-struct PaginationState: Equatable {
-    var totalRowCount: Int?         // Total rows in table (from COUNT(*))
-    var pageSize: Int               // Rows per page (passed from manager/coordinator)
-    var currentPage: Int = 1         // Current page number (1-based)
-    var currentOffset: Int = 0       // Current OFFSET for SQL query
-    var isLoading: Bool = false      // Loading indicator
-    var isApproximateRowCount: Bool = false  // True when totalRowCount is from fast estimate
-
-    /// Default page size constant (used when no explicit value is provided)
-    /// Note: For new tabs, callers should pass AppSettingsManager.shared.dataGrid.defaultPageSize
-    static let defaultPageSize = 1_000
-
-    init(
-        totalRowCount: Int? = nil,
-        pageSize: Int = PaginationState.defaultPageSize,
-        currentPage: Int = 1,
-        currentOffset: Int = 0,
-        isLoading: Bool = false
-    ) {
-        self.totalRowCount = totalRowCount
-        self.pageSize = pageSize
-        self.currentPage = currentPage
-        self.currentOffset = currentOffset
-        self.isLoading = isLoading
-    }
-
-    // MARK: - Computed Properties
-
-    /// Total number of pages
-    var totalPages: Int {
-        guard let total = totalRowCount, total > 0 else { return 1 }
-        return (total + pageSize - 1) / pageSize  // Ceiling division
-    }
-
-    /// Whether there is a next page available
-    var hasNextPage: Bool {
-        currentPage < totalPages
-    }
-
-    /// Whether there is a previous page available
-    var hasPreviousPage: Bool {
-        currentPage > 1
-    }
-
-    /// Starting row number for current page (1-based)
-    var rangeStart: Int {
-        currentOffset + 1
-    }
-
-    /// Ending row number for current page (1-based)
-    var rangeEnd: Int {
-        guard let total = totalRowCount else {
-            return currentOffset + pageSize
-        }
-        return min(currentOffset + pageSize, total)
-    }
-
-    // MARK: - Navigation Methods
-
-    /// Navigate to next page
-    mutating func goToNextPage() {
-        guard hasNextPage else { return }
-        currentPage += 1
-        currentOffset = (currentPage - 1) * pageSize
-    }
-
-    /// Navigate to previous page
-    mutating func goToPreviousPage() {
-        guard hasPreviousPage else { return }
-        currentPage -= 1
-        currentOffset = (currentPage - 1) * pageSize
-    }
-
-    /// Navigate to first page
-    mutating func goToFirstPage() {
-        currentPage = 1
-        currentOffset = 0
-    }
-
-    /// Navigate to last page
-    mutating func goToLastPage() {
-        currentPage = totalPages
-        currentOffset = (totalPages - 1) * pageSize
-    }
-
-    /// Navigate to specific page
-    mutating func goToPage(_ page: Int) {
-        guard page > 0 && page <= totalPages else { return }
-        currentPage = page
-        currentOffset = (page - 1) * pageSize
-    }
-
-    /// Reset pagination to first page
-    mutating func reset() {
-        currentPage = 1
-        currentOffset = 0
-        isLoading = false
-    }
-
-    /// Update page size (limit)
-    mutating func updatePageSize(_ newSize: Int) {
-        guard newSize > 0 else { return }
-        pageSize = newSize
-        // Recalculate current page based on current offset
-        currentPage = (currentOffset / pageSize) + 1
-    }
-
-    /// Update offset directly and recalculate page
-    mutating func updateOffset(_ newOffset: Int) {
-        guard newOffset >= 0 else { return }
-        currentOffset = newOffset
-        currentPage = (currentOffset / pageSize) + 1
-    }
-}
-
-/// Stores column layout (widths and order) within a tab session
-struct ColumnLayoutState: Equatable {
-    var columnWidths: [String: CGFloat] = [:]
-    var columnOrder: [String]?
-    var hiddenColumns: Set<String> = []
-}
-
-/// Reference-type wrapper for large result data.
-/// When QueryTab (a struct) is copied via CoW, only this 8-byte reference is copied
-/// instead of duplicating potentially large result arrays.
-final class RowBuffer {
-    var rows: [[String?]]
-    var columns: [String]
-    var columnTypes: [ColumnType]
-    var columnDefaults: [String: String?]
-    var columnForeignKeys: [String: ForeignKeyInfo]
-    var columnEnumValues: [String: [String]]
-    var columnNullable: [String: Bool]
-
-    init(
-        rows: [[String?]] = [],
-        columns: [String] = [],
-        columnTypes: [ColumnType] = [],
-        columnDefaults: [String: String?] = [:],
-        columnForeignKeys: [String: ForeignKeyInfo] = [:],
-        columnEnumValues: [String: [String]] = [:],
-        columnNullable: [String: Bool] = [:]
-    ) {
-        self.rows = rows
-        self.columns = columns
-        self.columnTypes = columnTypes
-        self.columnDefaults = columnDefaults
-        self.columnForeignKeys = columnForeignKeys
-        self.columnEnumValues = columnEnumValues
-        self.columnNullable = columnNullable
-    }
-
-    /// Create a deep copy of this buffer (used when explicit data duplication is needed)
-    func copy() -> RowBuffer {
-        RowBuffer(
-            rows: rows,
-            columns: columns,
-            columnTypes: columnTypes,
-            columnDefaults: columnDefaults,
-            columnForeignKeys: columnForeignKeys,
-            columnEnumValues: columnEnumValues,
-            columnNullable: columnNullable
-        )
-    }
-
-    /// Whether this buffer's row data has been evicted to save memory
-    private(set) var isEvicted: Bool = false
-
-    /// Evict row data to free memory. Column metadata is preserved.
-    func evict() {
-        guard !isEvicted else { return }
-        rows = []
-        isEvicted = true
-    }
-
-    /// Restore row data after eviction
-    func restore(rows newRows: [[String?]]) {
-        self.rows = newRows
-        isEvicted = false
-    }
-
-    deinit {
-        #if DEBUG
-        Logger(subsystem: "com.TablePro", category: "RowBuffer")
-            .debug("RowBuffer deallocated — columns: \(self.columns.count), evicted: \(self.isEvicted)")
-        #endif
-    }
-}
-
 /// Represents a single tab (query or table)
 struct QueryTab: Identifiable, Equatable {
     let id: UUID
@@ -572,201 +301,5 @@ struct QueryTab: Identifiable, Equatable {
             && lhs.isResultsCollapsed == rhs.isResultsCollapsed
             && lhs.resultSets.map(\.id) == rhs.resultSets.map(\.id)
             && lhs.activeResultSetId == rhs.activeResultSetId
-    }
-}
-
-/// Manager for query tabs
-@MainActor @Observable
-final class QueryTabManager {
-    var tabs: [QueryTab] = [] {
-        didSet { _tabIndexMapDirty = true }
-    }
-
-    var selectedTabId: UUID?
-
-    @ObservationIgnored private var _tabIndexMap: [UUID: Int] = [:]
-    @ObservationIgnored private var _tabIndexMapDirty = true
-
-    private func rebuildTabIndexMapIfNeeded() {
-        guard _tabIndexMapDirty else { return }
-        _tabIndexMap = Dictionary(uniqueKeysWithValues: tabs.enumerated().map { ($1.id, $0) })
-        _tabIndexMapDirty = false
-    }
-
-    var tabIds: [UUID] { tabs.map(\.id) }
-
-    var selectedTab: QueryTab? {
-        if let index = selectedTabIndex { return tabs[index] }
-        return selectedTabId == nil ? tabs.first : nil
-    }
-
-    var selectedTabIndex: Int? {
-        guard let id = selectedTabId else { return nil }
-        rebuildTabIndexMapIfNeeded()
-        return _tabIndexMap[id]
-    }
-
-    init() {
-        // Start with no tabs - shows empty state
-        tabs = []
-        selectedTabId = nil
-    }
-
-    // MARK: - Tab Management
-
-    func addTab(initialQuery: String? = nil, title: String? = nil, databaseName: String = "", sourceFileURL: URL? = nil) {
-        if let sourceFileURL,
-           let existingIndex = tabs.firstIndex(where: { $0.sourceFileURL == sourceFileURL }) {
-            if let query = initialQuery {
-                tabs[existingIndex].query = query
-            }
-            selectedTabId = tabs[existingIndex].id
-            return
-        }
-
-        let queryCount = tabs.count(where: { $0.tabType == .query })
-        let tabTitle = title ?? "Query \(queryCount + 1)"
-        var newTab = QueryTab(title: tabTitle, tabType: .query)
-
-        if let query = initialQuery {
-            newTab.query = query
-            newTab.hasUserInteraction = true
-        }
-
-        newTab.databaseName = databaseName
-        newTab.sourceFileURL = sourceFileURL
-        if sourceFileURL != nil {
-            newTab.savedFileContent = newTab.query
-        }
-        tabs.append(newTab)
-        selectedTabId = newTab.id
-    }
-
-    func addTableTab(
-        tableName: String,
-        databaseType: DatabaseType = .mysql,
-        databaseName: String = "",
-        quoteIdentifier: ((String) -> String)? = nil
-    ) {
-        // Check if table tab already exists (match on databaseName)
-        if let existingTab = tabs.first(where: {
-            $0.tabType == .table && $0.tableName == tableName && $0.databaseName == databaseName
-        }) {
-            selectedTabId = existingTab.id
-            return
-        }
-
-        let pageSize = AppSettingsManager.shared.dataGrid.defaultPageSize
-        let query = QueryTab.buildBaseTableQuery(
-            tableName: tableName, databaseType: databaseType, quoteIdentifier: quoteIdentifier
-        )
-        var newTab = QueryTab(
-            title: tableName,
-            query: query,
-            tabType: .table,
-            tableName: tableName
-        )
-        newTab.pagination = PaginationState(pageSize: pageSize)
-        newTab.databaseName = databaseName
-        tabs.append(newTab)
-        selectedTabId = newTab.id
-    }
-
-    func addCreateTableTab(databaseName: String = "") {
-        let tabTitle = String(localized: "Create Table")
-        var newTab = QueryTab(title: tabTitle, tabType: .createTable)
-        newTab.databaseName = databaseName
-        newTab.isEditable = false
-        newTab.hasUserInteraction = true
-        tabs.append(newTab)
-        selectedTabId = newTab.id
-    }
-
-    func addPreviewTableTab(
-        tableName: String,
-        databaseType: DatabaseType = .mysql,
-        databaseName: String = "",
-        quoteIdentifier: ((String) -> String)? = nil
-    ) {
-        let pageSize = AppSettingsManager.shared.dataGrid.defaultPageSize
-        let query = QueryTab.buildBaseTableQuery(
-            tableName: tableName, databaseType: databaseType, quoteIdentifier: quoteIdentifier
-        )
-        var newTab = QueryTab(
-            title: tableName,
-            query: query,
-            tabType: .table,
-            tableName: tableName
-        )
-        newTab.pagination = PaginationState(pageSize: pageSize)
-        newTab.databaseName = databaseName
-        newTab.isPreview = true
-        tabs.append(newTab)
-        selectedTabId = newTab.id
-    }
-
-    /// Replace the currently selected tab's content with a new table.
-    /// - Returns: `true` if the replacement happened (caller should run the query),
-    ///   `false` if there is no selected tab.
-    @discardableResult
-    func replaceTabContent(
-        tableName: String, databaseType: DatabaseType = .mysql,
-        isView: Bool = false, databaseName: String = "",
-        schemaName: String? = nil, isPreview: Bool = false,
-        quoteIdentifier: ((String) -> String)? = nil
-    ) -> Bool {
-        guard let selectedId = selectedTabId,
-              let selectedIndex = tabs.firstIndex(where: { $0.id == selectedId })
-        else {
-            return false
-        }
-
-        let query = QueryTab.buildBaseTableQuery(
-            tableName: tableName,
-            databaseType: databaseType,
-            schemaName: schemaName,
-            quoteIdentifier: quoteIdentifier
-        )
-        let pageSize = AppSettingsManager.shared.dataGrid.defaultPageSize
-
-        // Build locally and write back once to avoid 14 CoW copies (UI-11).
-        var tab = tabs[selectedIndex]
-        tab.rowBuffer = RowBuffer()
-        tab.title = tableName
-        tab.tableName = tableName
-        tab.query = query
-        tab.resultVersion += 1
-        tab.executionTime = nil
-        tab.statusMessage = nil
-        tab.errorMessage = nil
-        tab.lastExecutedAt = nil
-        tab.showStructure = false
-        tab.sortState = SortState()
-        tab.selectedRowIndices = []
-        tab.pendingChanges = TabPendingChanges()
-        tab.hasUserInteraction = false
-        tab.isView = isView
-        tab.isEditable = !isView
-        tab.filterState = TabFilterState()
-        tab.columnLayout = ColumnLayoutState()
-        tab.pagination = PaginationState(pageSize: pageSize)
-        tab.databaseName = databaseName
-        tab.schemaName = schemaName
-        tab.isPreview = isPreview
-        tabs[selectedIndex] = tab
-        return true
-    }
-
-    func updateTab(_ tab: QueryTab) {
-        if let index = tabs.firstIndex(where: { $0.id == tab.id }) {
-            tabs[index] = tab
-        }
-    }
-
-    deinit {
-        #if DEBUG
-        Logger(subsystem: "com.TablePro", category: "QueryTabManager")
-            .debug("QueryTabManager deallocated")
-        #endif
     }
 }
